+++ src/main/java/cn/haitaoss/tinyspringmvc/framework/handlerAdapter/AnnotationHandlerAdapter.java
@@ -2,6 +2,7 @@ package cn.haitaoss.tinyspringmvc.framework.handlerAdapter;
 
 import cn.haitaoss.tinyioc.context.ApplicationContext;
 import cn.haitaoss.tinyspringmvc.framework.handlerMapping.RequestMappingHandler;
+import cn.haitaoss.tinyspringmvc.framework.modelAndView.Model;
 import cn.haitaoss.tinyspringmvc.framework.modelAndView.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
@@ -29,10 +30,21 @@ public class AnnotationHandlerAdapter extends AbstractHandlerAdapter {
     @Override
     public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
         RequestMappingHandler rmHandler = (RequestMappingHandler) handler;
+        Model model = new Model();
         // 获取方法的执行参数的值
-        Object[] args = ArgumentResolverUtil.resolveRequestParam(request, rmHandler.getMethod());
+        Object[] args = ArgumentResolverUtil.resolveRequestParam(request, rmHandler.getMethod(), model);
         // 执行方法
         Object obj = rmHandler.getMethod().invoke(rmHandler.getBean(), args);
-        return new ModelAndView(obj);
+
+        if (obj instanceof ModelAndView) {
+            return (ModelAndView) obj;
+        }
+        ModelAndView mv = new ModelAndView();
+        if (obj instanceof String) {
+            mv.setModel(model);
+            mv.setView((String) obj);
+        }
+        return mv;
+
     }
 }