+++ src/main/java/cn/haitaoss/tinyspringmvc/framework/handlerAdapter/ArgumentResolverUtil.java
@@ -1,6 +1,8 @@
 package cn.haitaoss.tinyspringmvc.framework.handlerAdapter;
 
 import cn.haitaoss.tinyioc.beans.converter.ConverterFactory;
+import cn.haitaoss.tinyspringmvc.framework.modelAndView.Model;
+import cn.haitaoss.tinyspringmvc.framework.modelAndView.ModelAndView;
 import org.apache.commons.beanutils.BeanUtils;
 
 import javax.servlet.http.HttpServletRequest;
@@ -17,7 +19,7 @@ import java.util.Map;
  *
  */
 public class ArgumentResolverUtil {
-    public static Object[] resolveRequestParam(HttpServletRequest request, Method method) throws Exception {
+    public static Object[] resolveRequestParam(HttpServletRequest request, Method method, Model model) throws Exception {
         // 获取方法的参数值
         Parameter[] parameters = method.getParameters();
         Object[] args = new Object[parameters.length];
@@ -40,27 +42,42 @@ public class ArgumentResolverUtil {
             }
             // 处理后的request键值对
             argMap.put(paramName, paramValue);
+            // 将这些参数保存到 model中
+            model.put(paramName, paramValue);
         }
 
         for (int i = 0; i < parameters.length; i++) {
             Parameter parameter = parameters[i];
+            Class<?> parameterType = parameter.getType();
+            // 如果形参中有Model类，则创建一个参数
+            if (parameterType == Model.class) {
+                args[i] = model;
+                continue;
+            }
+            if (parameterType == ModelAndView.class) {
+                args[i] = new ModelAndView(model);
+                continue;
+            }
+            if (parameterType == HttpServletRequest.class) {
+                args[i] = request;
+                continue;
+            }
             if (argMap.containsKey(parameter.getName())) {
                 String value = argMap.get(parameter.getName());
-                Type type = parameter.getType();
+                Type type = parameterType;
                 // 字符串类型，直接赋值
                 if (type == String.class)
                     args[i] = value;
                 else
                     // 复杂参数类型 通过converter解析在赋值
-                    args[i] = ConverterFactory.getConverterMap().get(parameter.getType()).parse(value);
+                    args[i] = ConverterFactory.getConverterMap().get(parameterType).parse(value);
             } else {
                 // 没有的参数，就通过反射创建实例
-                Type type = parameter.getType();
-                Object bean = ((Class) type).newInstance();
+                Object bean = parameterType.newInstance();
                 try {
                     // 将map里面的属性 拷贝到新创建的实例中
                     BeanUtils.populate(bean, argMap);
-                    args[i] = ((Class) type).cast(bean);
+                    args[i] = parameterType.cast(bean);
                 } catch (Exception e) {
                     args[i] = null;
                 }