+++ src/main/java/cn/haitaoss/tinyspringmvc/framework/modelAndView/Model.java
@@ -1,0 +1,13 @@
+package cn.haitaoss.tinyspringmvc.framework.modelAndView;
+
+import java.util.LinkedHashMap;
+
+/**
+ * @author haitao.chen
+ * email haitaoss@aliyun.com
+ * date 2021-04-25 16:18
+ *
+ */
+public class Model extends LinkedHashMap<String,Object> {
+
+}