+++ src/main/java/cn/haitaoss/tinyspringmvc/framework/modelAndView/ModelAndView.java
@@ -7,17 +7,34 @@ package cn.haitaoss.tinyspringmvc.framework.modelAndView;
  *
  */
 public class ModelAndView {
-    private Object bean;
+    private Model model;
+    private String view;
 
-    public ModelAndView(Object bean) {
-        this.bean = bean;
+    public ModelAndView() {
     }
 
-    public Object getBean() {
-        return bean;
+    public ModelAndView(Model model) {
+        this.model = model;
     }
 
-    public void setBean(Object bean) {
-        this.bean = bean;
+    public ModelAndView(Model model, String view) {
+        this.model = model;
+        this.view = view;
+    }
+
+    public Model getModel() {
+        return model;
+    }
+
+    public void setModel(Model model) {
+        this.model = model;
+    }
+
+    public String getView() {
+        return view;
+    }
+
+    public void setView(String view) {
+        this.view = view;
     }
 }