+++ src/main/java/cn/haitaoss/tinyspringmvc/framework/view/View.java
@@ -1,0 +1,17 @@
+package cn.haitaoss.tinyspringmvc.framework.view;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.Map;
+
+/**
+ * @author haitao.chen
+ * email haitaoss@aliyun.com
+ * date 2021-04-25 16:20
+ *
+ */
+public interface View {
+    void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException;
+}